% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_data_html.R
\name{embed_data_html}
\alias{embed_data_html}
\title{Embed R data in HTML file}
\source{
\url{http://bayesfactor.blogspot.com.es/2014/09/embedding-rdata-files-in-rmarkdown.html}
}
\usage{
embed_data_html(dataset, textHTML = "Click here to download the data")
}
\arguments{
\item{dataset}{Character (required). Name of the dataset (RData or rda file) to be embedded in the HTML file (skipping the extension, i.e. rda or Rdata). 
Must be included within quotes.}

\item{textHTML}{Character (optional). Text to be shown in the download link.}
}
\description{
This function generates HTML code to encode an RData/rda file within the HTML file.
}
\author{
Richard Morey, slightly modified by F. Rodriguez-Sanchez.
}
