% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_certificates.R
\name{generate_certificates}
\alias{generate_certificates}
\title{Generate certificates}
\usage{
generate_certificates(names = NULL, rmd = NULL, outdir = "certificates")
}
\arguments{
\item{names}{A character vector of people names}

\item{rmd}{Path to an Rmarkdown parameterized document.}

\item{outdir}{Output directory where individual certificates will be saved}
}
\value{
One or many PDF files with individual certificates
}
\description{
Automatically generate personal PDF certificates from an Rmarkdown 
parameterized template and a vector of people names.
}
