% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdTemplates.R
\docType{package}
\name{rmdTemplates}
\alias{rmdTemplates}
\alias{rmdTemplates-package}
\title{A collection of Rmarkdown templates}
\description{
A collection of Rmarkdown templates. Include helpful defaults for
knitr options, as well as support for citations and different bibliography styles.
}
\details{
After installing the package, click on `create new Rmarkdown document` in Rstudio
to have access to the templates. 
See \url{http://rmarkdown.rstudio.com/developer_document_templates.html} for 
more information.
}
\author{
F. Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com}
}
\keyword{package}

