% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_rmd_html.R
\name{embed_rmd_html}
\alias{embed_rmd_html}
\title{Embed Rmarkdown source code in the output HTML file.}
\source{
\url{http://rpubs.com/ramnathv/including_rmd_source}
}
\usage{
embed_rmd_html(file)
}
\arguments{
\item{file}{Character. Name of the Rmarkdown source file (within quotes), e.g. 'myreport.Rmd'}
}
\description{
This function generates HTML code to encode the Rmarkdown source code within the output HTML file.
}
\author{
Ramnath Vaidyanathan, slightly modified by F. Rodriguez-Sanchez.
}
